requirements.txt
numpy
python-opencv
Pillow
streamlit
fastapi
uvicorn
python-multipart
torch
torchvision
ultralytics
datasets